Improve song formatting
Format the rich presence state to include the artist and album name,
instead of the artist and track name. The track name is already included
in the details field, so its presence in the state field is superfluous.

Make the small text more succinct.

Furthermore, refactor the code paths and rename variables to be less
confusing.